一旦全部完成させますね

+++ app/src/main/java/com/example/calendar/data/entity/Tag.kt
@@ -4,7 +4,7 @@ package com.example.calendar.data.entity
  * ER図の Tag テーブルに対応するデータクラス
  */
 data class Tag(
-    val tagId: Int = 0,         // PK
+    val tagId: Long = 0,         // PK
     val name: String,          
     val icon: String? = null,
     val color: Int,
